Documents/2024 Interview projects/JBT-Test/jbt-test/requirements.txt
Flask==3.0.1
PyJWT==2.8.0
geopy==2.4.1
pytest==8.0.1